Substantially revamped options with OptsBase.plot_ calls to change the plotting call, cosmetic grouping and ordering of OptsBase methods.
Added autolegend option to automatically generate a legend entry if
none was supplied.

You can now do a particular type of plot (on the sp context object or
the SpecialAx that a call to it returns) by calling that attribute of
the object instead of just the object itself. Or, you can do
sp.plot_xxx to have the default call made into xxx instead of 'plot'.