RELEASE 0.9.5
This should be the last Beta release. The only thing left to do before
considering this a production-level package is to test on Python 3.x
and fix any incompatibilities.

This commit changes the end date of the copyright notice to 2020 and
adds the h2 and w2 keywords to Plotter and SubPlotter for doubling
height and width of selected subplot row(s) or column(s).

Since RELEASE 0.9.4
-------------------------------------------------------------
Included new test subpackage.

Grid includes faint grey dotted lines for minor ticks.

Fixed bug and cleaned up copy-paste code smell for module-level
functions xy and xyc.

Improved subplot spacing, fixing bug where ylabel spacing needs were
being ignored if there was just a single subplot or column of
subplots.

Fixed bug where axvline position was not being scaled along with
x-axis when timex option set.

Added 'start' and 'done' methods to Plotter as an alternative way of
using it instead of calling in context.

Adds the h2 and w2 keywords to Plotter and SubPlotter for doubling
height and width of selected subplot row(s) or column(s).