RELEASE 0.9.6
This should be the last Beta release. The only thing left to do before
considering this a production-level package is to test on Python 3.x
and fix any incompatibilities.

I did need to do a minor release because of some small improvements
that are used in my latest example covid19.py of my 'ade' package.